+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### Fixed bugs:
 * Fix macOS application bundle
+* Handle empty scoreboard correctly
 
 ## Release 0.7.5 (2020-04-24)